Relay-Service/requirements.txt
python-dotenv
greenlet==3.2.4
pika==1.3.2
psycopg2-binary==2.9.11
SQLAlchemy==2.0.44
typing_extensions==4.15.0
